[RatingInput#1676] Move closestTo to an extension.
